requirements.txt
nicegui
PyYAML
clickhouse-connect
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc
pytest
pytest-cov
